+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Upcoming release
 
+### Bugfixes
+
+- `no-implicit-close` now provides more context when closed by a sibling.
+- `close-order` no longer reports error for implicitly closed elements.
+
 ## 0.11.0 (2018-09-23)
 
 ### Breaking changes